Code clean-up: Move ScopeComparator class
This class is not used by ScopeTreeData anymore since in order to sort
nodes, we need to calculate the depth as well (required by nattable).
