+++ edu.rice.cs.hpcdata.app/src/edu/rice/cs/hpcdata/app/PrintData.java
@@ -15,7 +15,6 @@ import edu.rice.cs.hpcdata.experiment.metric.MetricValue;
 import edu.rice.cs.hpcdata.experiment.scope.RootScope;
 import edu.rice.cs.hpcdata.experiment.scope.RootScopeType;
 import edu.rice.cs.hpcdata.experiment.scope.Scope;
-import edu.rice.cs.hpcdata.util.ScopeComparator;
 import edu.rice.cs.hpcdata.util.Util;
